+++ RELEASE_NOTES.md
@@ -1,11 +1,10 @@
 ## 0.1.3 (next release)
 
-- __New feature__: One can now pass custom arguments to NUnit via the key `Bud.CSharp.NUnitTestTargetPlugin.NUnitArgumentsKey` in the `Cs.Test()` build target. For example:
+- __New feature__: One can now pass custom arguments to NUnit via the key `Bud.CSharp.NUnitPlugin.NUnitArgumentsKey` in the `Cs.Test()` build target. For example:
 
     ```language-csharp
     new Project("Foo.Project", baseDir, Cs.Test(
-        Cs.Dependency("Bud.Test"),
-        NUnitTestTargetPlugin.NUnitArgumentsKey.Modify(list => list.Add("/noshadow"))
+      NUnitPlugin.NUnitArgs.Modify(list => list.Add("/noshadow"))
     ))
     ```